Refactor into separate files + add xmldocs + rename some methods
